Created BB and BT for Zebra + started prefab
Created the blackboard and behaviour tree for the Zebra.
Started filling in this prefab, adding the tree with blackboard.

Created a MoveToTarget node, this will move the target towards
This is a shared action for every tree and uses the animal position and targetPosition.

Created more materials.

Created base prefabs for the water and the grass.

Modified the AnimalStats to include speed.