Created Herd + Conditionals + Partial Sequencer
Created the Zebra Herd.
This spawns a number of zebras.
It also finds all the grass and water nearby and stores it in their memory.

Created tags for Grass and Water.

Created a Conditional to check if the animal is still alive.

Created a Partial Sequencer (I actually already made this but mistakingly called it just a Sequencer).
The Sequencer has been fixed.