Made improvements -> removed redundant stuff
Improved the BaseAnimalStats:
The animal will stay thirsty/hungry untill they have satisfied their needs.
This removes the need of these useless scripts that determine of they should eat/drink.

Created Wandering.
If the zebras aren't hungry or thirsty, they will start wandering towards a random point in the map.
This point is randomised every 5 seconds in the herd script.

Fixed things so that the scripts for either BT or FSM won't be active when it doesn't need to be.